Move CAA types selector to a separate file
